requirements.txt
# Django core
Django==5.2

# Base de datos
mysqlclient==2.2.7

# Websockets y tiempo real
channels==4.2.2
channels-redis==4.2.1

# Redis (capa de cache y pub/sub)
redis==5.2.1

# Celery (tareas asíncronas y programadas)
celery==5.4.0
django-celery-beat==2.7.0

# Configuración por entorno
python-dotenv==1.1.0

# Autocompletado en formularios
django-autocomplete-light==3.12.1

# MPTT (estructuras jerárquicas)
django-mptt==0.17.0

# Django REST Framework
djangorestframework==3.16.0

# CORS para APIs
django-cors-headers==4.7.0

# Manejo de íconos en Django
django-icons==25.1

# Archivos estáticos en producción
whitenoise==6.9.0

# PDF
PyPDF2==3.0.1
xhtml2pdf==0.2.17

# Excel
openpyxl==3.1.5
pandas==2.3.2

# HTTP
requests==2.32.3
urllib3==2.4.0
chardet==5.2.0
certifi==2025.1.31
idna==3.10

# Seguridad y JWT
cryptography==44.0.2
pyOpenSSL==25.0.0
PyJWT==2.10.1

# Utilidades
PyYAML==6.0.2
pytz==2025.2

# Jinja (por si usas templates personalizados o generación HTML externa)
MarkupSafe==3.0.2
Jinja2==3.1.6

# Servidores para producción
gunicorn==23.0.0
uvicorn==0.34.2

django-select2==8.4.0

# Dbug para vs code
debugpy==1.8.16